refactor(slack-runtime): split command/render helpers (#1587)
