+++ crates/tau-slack-runtime/src/slack_runtime.rs
@@ -12,7 +12,7 @@ use futures_util::{SinkExt, StreamExt};
 use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 use tau_agent_core::{Agent, AgentConfig, AgentEvent};
-use tau_ai::{LlmClient, Message, MessageRole};
+use tau_ai::LlmClient;
 use tokio::sync::watch;
 use tokio_tungstenite::{connect_async, tungstenite::Message as WsMessage};
 
@@ -65,9 +65,20 @@ pub struct SlackBridgeRuntimeConfig {
 }
 
 mod slack_api_client;
+mod slack_command_helpers;
+mod slack_render_helpers;
 mod slack_state_store;
 
 use slack_api_client::{SlackApiClient, SlackPostedMessage};
+use slack_command_helpers::{
+    parse_slack_command, rbac_action_for_slack_command, render_slack_command_response,
+    slack_command_usage,
+};
+use slack_render_helpers::{
+    collect_assistant_reply, normalize_artifact_retention_days, prompt_status_label,
+    render_event_prompt, render_slack_artifact_markdown, render_slack_response,
+    render_slack_run_error_message, slack_metadata_marker,
+};
 use slack_state_store::{JsonlEventLog, SlackBridgeStateStore};
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
@@ -1528,363 +1539,6 @@ fn event_is_stale(event: &SlackBridgeEvent, max_event_age_seconds: u64, now_unix
     }
     let max_age_ms = max_event_age_seconds.saturating_mul(1000);
     now_unix_ms.saturating_sub(event.occurred_unix_ms) > max_age_ms
-}
-
-fn normalize_slack_message_text(event: &SlackBridgeEvent, bot_user_id: &str) -> String {
-    let mut message_text = event.text.trim().to_string();
-    if event.kind == SlackBridgeEventKind::AppMention {
-        let mention = format!("<@{bot_user_id}>");
-        message_text = message_text.replace(&mention, "");
-        message_text = message_text.trim().to_string();
-    }
-    message_text
-}
-
-fn render_event_prompt(
-    event: &SlackBridgeEvent,
-    bot_user_id: &str,
-    downloaded_files: &[DownloadedSlackFile],
-) -> String {
-    let message_text = normalize_slack_message_text(event, bot_user_id);
-
-    let mut prompt = format!(
-        "You are responding as Tau inside Slack.\nChannel: {}\nUser: <@{}>\nEvent kind: {}\nMessage ts: {}\n\nUser message:\n{}",
-        event.channel_id,
-        event.user_id,
-        event.kind.as_str(),
-        event.ts,
-        if message_text.is_empty() {
-            "(empty message)"
-        } else {
-            &message_text
-        }
-    );
-
-    if !downloaded_files.is_empty() {
-        prompt.push_str("\n\nDownloaded attachments:\n");
-        for file in downloaded_files {
-            prompt.push_str(&format!(
-                "- id={} name={} path={} mimetype={} size={}\n",
-                file.id,
-                file.original_name,
-                file.path.display(),
-                file.mimetype
-                    .clone()
-                    .unwrap_or_else(|| "unknown".to_string()),
-                file.size.unwrap_or(0)
-            ));
-        }
-    }
-
-    prompt.push_str("\nProvide a direct, concise Slack-ready response.");
-    prompt
-}
-
-fn slack_command_usage() -> String {
-    [
-        "Supported `/tau` commands:",
-        "- `/tau help`",
-        "- `/tau status`",
-        "- `/tau health`",
-        "- `/tau stop`",
-        "- `/tau artifacts [purge|run <run_id>|show <artifact_id>]`",
-        "- `/tau canvas <create|update|show|export|import> ...`",
-    ]
-    .join("\n")
-}
-
-fn rbac_action_for_slack_command(command: Option<&SlackCommand>) -> String {
-    match command {
-        Some(SlackCommand::Help) => "command:/tau-help".to_string(),
-        Some(SlackCommand::Status) => "command:/tau-status".to_string(),
-        Some(SlackCommand::Health) => "command:/tau-health".to_string(),
-        Some(SlackCommand::Stop) => "command:/tau-stop".to_string(),
-        Some(SlackCommand::Artifacts { .. }) => "command:/tau-artifacts".to_string(),
-        Some(SlackCommand::ArtifactShow { .. }) => "command:/tau-artifacts-show".to_string(),
-        Some(SlackCommand::Canvas { .. }) => "command:/tau-canvas".to_string(),
-        Some(SlackCommand::Invalid { .. }) => "command:/tau-invalid".to_string(),
-        None => "command:/tau-run".to_string(),
-    }
-}
-
-fn parse_slack_command(event: &SlackBridgeEvent, bot_user_id: &str) -> Option<SlackCommand> {
-    let normalized = normalize_slack_message_text(event, bot_user_id);
-    let trimmed = normalized.trim();
-    let mut pieces = trimmed.split_whitespace();
-    let command_prefix = pieces.next()?;
-    if command_prefix != "/tau" {
-        return None;
-    }
-
-    let args = trimmed[command_prefix.len()..].trim();
-    if args.is_empty() {
-        return Some(SlackCommand::Invalid {
-            message: slack_command_usage(),
-        });
-    }
-    let mut parts = args.splitn(2, char::is_whitespace);
-    let command = parts.next().unwrap_or_default();
-    let remainder = parts.next().unwrap_or_default().trim();
-    let parsed = match command {
-        "help" => {
-            if remainder.is_empty() {
-                SlackCommand::Help
-            } else {
-                SlackCommand::Invalid {
-                    message: "Usage: /tau help".to_string(),
-                }
-            }
-        }
-        "status" => {
-            if remainder.is_empty() {
-                SlackCommand::Status
-            } else {
-                SlackCommand::Invalid {
-                    message: "Usage: /tau status".to_string(),
-                }
-            }
-        }
-        "health" => {
-            if remainder.is_empty() {
-                SlackCommand::Health
-            } else {
-                SlackCommand::Invalid {
-                    message: "Usage: /tau health".to_string(),
-                }
-            }
-        }
-        "stop" => {
-            if remainder.is_empty() {
-                SlackCommand::Stop
-            } else {
-                SlackCommand::Invalid {
-                    message: "Usage: /tau stop".to_string(),
-                }
-            }
-        }
-        "artifacts" => {
-            if remainder.is_empty() {
-                SlackCommand::Artifacts {
-                    purge: false,
-                    run_id: None,
-                }
-            } else if remainder == "purge" {
-                SlackCommand::Artifacts {
-                    purge: true,
-                    run_id: None,
-                }
-            } else {
-                let mut artifact_args = remainder.split_whitespace();
-                match (
-                    artifact_args.next(),
-                    artifact_args.next(),
-                    artifact_args.next(),
-                ) {
-                    (Some("run"), Some(run_id), None) => SlackCommand::Artifacts {
-                        purge: false,
-                        run_id: Some(run_id.to_string()),
-                    },
-                    (Some("show"), Some(artifact_id), None) => SlackCommand::ArtifactShow {
-                        artifact_id: artifact_id.to_string(),
-                    },
-                    _ => SlackCommand::Invalid {
-                        message: "Usage: /tau artifacts [purge|run <run_id>|show <artifact_id>]"
-                            .to_string(),
-                    },
-                }
-            }
-        }
-        "canvas" => {
-            if remainder.is_empty() {
-                SlackCommand::Invalid {
-                    message: "Usage: /tau canvas <create|update|show|export|import> ..."
-                        .to_string(),
-                }
-            } else {
-                SlackCommand::Canvas {
-                    args: remainder.to_string(),
-                }
-            }
-        }
-        _ => SlackCommand::Invalid {
-            message: format!(
-                "Unknown command `{}`.\n\n{}",
-                command,
-                slack_command_usage()
-            ),
-        },
-    };
-    Some(parsed)
-}
-
-fn render_slack_response(
-    event: &SlackBridgeEvent,
-    run: &PromptRunReport,
-    detail_thread_output: bool,
-    detail_threshold_chars: usize,
-) -> (String, Option<String>) {
-    let reply = run.assistant_reply.trim();
-    let base_reply = if reply.is_empty() {
-        "I couldn't generate a textual response for this Slack event."
-    } else {
-        reply
-    };
-    let usage = &run.usage;
-    let status = format!("{:?}", run.status).to_lowercase();
-
-    let mut summary_body = base_reply.to_string();
-    let mut detail_body = None;
-
-    if detail_thread_output && base_reply.chars().count() > detail_threshold_chars.max(1) {
-        let summary = truncate_for_slack(base_reply, detail_threshold_chars.max(1));
-        summary_body = format!("{}\n\n(full response posted in this thread)", summary);
-        detail_body = Some(base_reply.to_string());
-    }
-
-    summary_body.push_str("\n\n---\n");
-    summary_body.push_str(&format!(
-        "{}\nTau run {} | status {} | model {} | tokens {}/{}/{}",
-        slack_metadata_marker(&event.key),
-        run.run_id,
-        status,
-        run.model,
-        usage.input_tokens,
-        usage.output_tokens,
-        usage.total_tokens
-    ));
-    summary_body.push_str(&format!(
-        "\nartifact {} | sha256 {} | bytes {}",
-        run.artifact.relative_path, run.artifact.checksum_sha256, run.artifact.bytes
-    ));
-
-    if !run.downloaded_files.is_empty() {
-        summary_body.push_str("\nattachments downloaded:");
-        for file in &run.downloaded_files {
-            summary_body.push_str(&format!(
-                "\n- {} ({})",
-                file.original_name,
-                file.path.display()
-            ));
-        }
-    }
-
-    (truncate_for_slack(&summary_body, 38_000), detail_body)
-}
-
-fn render_slack_run_error_message(
-    event: &SlackBridgeEvent,
-    run_id: &str,
-    error: &anyhow::Error,
-) -> String {
-    truncate_for_slack(
-        &format!(
-            "Tau run {} failed for event {}.\n\nError: {}\n\n---\n{}",
-            run_id,
-            event.key,
-            truncate_for_error(&error.to_string(), 600),
-            slack_metadata_marker(&event.key),
-        ),
-        38_000,
-    )
-}
-
-fn render_slack_command_response(
-    event: &SlackBridgeEvent,
-    command_name: &str,
-    status: &str,
-    message: &str,
-) -> String {
-    let mut content = if message.trim().is_empty() {
-        "Tau command response.".to_string()
-    } else {
-        message.trim().to_string()
-    };
-    let command = if command_name.trim().is_empty() {
-        "unknown"
-    } else {
-        command_name.trim()
-    };
-    let status_label = if status.trim().is_empty() {
-        "reported"
-    } else {
-        status.trim()
-    };
-    content.push_str("\n\n---\n");
-    content.push_str(&format!(
-        "{}\nTau command `{}` | status `{}`",
-        slack_metadata_marker(&event.key),
-        command,
-        status_label
-    ));
-    truncate_for_slack(&content, 38_000)
-}
-
-fn slack_metadata_marker(event_key: &str) -> String {
-    format!("{SLACK_METADATA_MARKER_PREFIX}{event_key}{SLACK_METADATA_MARKER_SUFFIX}")
-}
-
-fn render_slack_artifact_markdown(
-    event: &SlackBridgeEvent,
-    run_id: &str,
-    status: PromptRunStatus,
-    assistant_reply: &str,
-    downloaded_files: &[DownloadedSlackFile],
-) -> String {
-    let mut lines = vec![
-        "# Tau Slack Artifact".to_string(),
-        format!("channel_id: {}", event.channel_id),
-        format!("event_key: {}", event.key),
-        format!("event_kind: {}", event.kind.as_str()),
-        format!("run_id: {}", run_id),
-        format!("status: {}", prompt_status_label(status)),
-    ];
-    if downloaded_files.is_empty() {
-        lines.push("attachments: none".to_string());
-    } else {
-        lines.push(format!("attachments: {}", downloaded_files.len()));
-        for file in downloaded_files {
-            lines.push(format!(
-                "- {} ({})",
-                file.original_name,
-                file.path.display()
-            ));
-        }
-    }
-    lines.push(String::new());
-    lines.push("## Assistant Reply".to_string());
-    lines.push(assistant_reply.trim().to_string());
-    lines.join("\n")
-}
-
-fn normalize_artifact_retention_days(days: u64) -> Option<u64> {
-    if days == 0 {
-        None
-    } else {
-        Some(days)
-    }
-}
-
-fn collect_assistant_reply(messages: &[Message]) -> String {
-    let content = messages
-        .iter()
-        .filter(|message| message.role == MessageRole::Assistant)
-        .map(Message::text_content)
-        .filter(|text| !text.trim().is_empty())
-        .collect::<Vec<_>>()
-        .join("\n\n");
-    if content.trim().is_empty() {
-        "I couldn't generate a textual response for this event.".to_string()
-    } else {
-        content
-    }
-}
-
-fn prompt_status_label(status: PromptRunStatus) -> &'static str {
-    match status {
-        PromptRunStatus::Completed => "completed",
-        PromptRunStatus::Cancelled => "cancelled",
-        PromptRunStatus::TimedOut => "timed_out",
-    }
 }
 
 #[cfg(test)]
